combined JWT auth for frontend and oauth2 support
