+++ src/db.ts
@@ -1,26 +1,17 @@
-const mysql = require('mysql2/promise');
-const dbConfiguration = require('./dbConfig');
+import mysql from 'mysql2/promise';
+import { mysqlConfig } from './config';
 
-async function createConnection() {
-    const { host, port, database, user, password } = dbConfiguration;
-    try {
-        const connection = await mysql.createConnection({
-            host: dbConfiguration.host,
-            port: dbConfiguration.port,
-            database: dbConfiguration.database,
-            user: dbConfiguration.user,
-            password: dbConfiguration.password
-        });
+// TODO: UPdate this to use the datasource defined and the connection pool
 
-        await connection.connect(); //Connect to database
+export async function createConnection() {
+  try {
+    const connection = await mysql.createConnection(mysqlConfig);
+    await connection.connect(); //Connect to database
 
-        console.log('Connected to the database as ID', connection.threadId);
-        return connection;
-    } catch (err) {
-        console.error('Error connecting to the database: ', err.stack);
-        throw err;
-    }
+    console.log('Connected to the database as ID', connection.threadId);
+    return connection;
+  } catch (err) {
+    console.error('Error connecting to the database: ', err.stack);
+    throw err;
+  }
 }
-
-
-module.exports = createConnection;
