+++ src/index.ts
@@ -10,13 +10,13 @@ import { healthcheck } from './controllers/healthcheck';
 import { attachApolloServer } from './middleware/express';
 import router from './router';
 
+// TODO: Make this configurable and pass in as ENV variable
 const PORT = 4000;
 
 // Required logic for integrating with Express
 const app = express();
 // Our httpServer handles incoming requests to our Express app.
 const httpServer = http.createServer(app);
-
 
 const apolloServer = new ApolloServer(serverConfig(logger, httpServer));
 
@@ -27,20 +27,19 @@ const startServer = async () => {
   // Healthcheck endpoint (declare this BEFORE Oauth2 and CORS definition due to AWS ALB limitations)
   app.get('/up', (_request, response) => healthcheck(apolloServer, response, logger));
 
-  // Initialize the OAuth2 server
-  initOAuthServer(app);
-
   // Express middleware
-
   app.use(
     cors(),
     express.urlencoded({ extended: false }),
     express.json({ limit: '50mb' }),
   )
 
+  // Attach Apollo server to all of the GraphQL calls
   app.use('/graphql', attachApolloServer(apolloServer, cache, logger))
 
+  // Pass off to the Router for other handling
   app.use('/', router);
+
   httpServer.listen({ port: 4000 }, () => {
     console.log(`Server running on port ${PORT}`);
     console.log(`GraphQL endpoint: http://localhost:${PORT}/graphql`)
